Unify function to submit partial config
