+++ src/plugins/staticHosting/plugin.ts
@@ -2,8 +2,9 @@ import * as core from '@actions/core';
 import fs from 'fs';
 
 import { StaticHostingPlugin } from '../../schemas/stoatConfigSchema';
-import { GithubActionRun } from '../../types';
-import { createSignedUrl, submitPartialConfig, uploadDirectory } from './helpers';
+import { GithubActionRun, UploadStaticHostingRequest } from '../../types';
+import { submitPartialConfig } from '../helpers';
+import { createSignedUrl, uploadDirectory } from './helpers';
 
 const runStaticHostingPlugin = async (
   taskId: string,
@@ -26,7 +27,7 @@ const runStaticHostingPlugin = async (
     ghRepo: repo,
     ghSha,
     ghToken,
-    taskId: taskId
+    taskId
   });
 
   // upload directory
@@ -34,7 +35,14 @@ const runStaticHostingPlugin = async (
   await uploadDirectory(signedUrl, fields, pathToUpload, objectPath);
 
   // submit partial config
-  await submitPartialConfig(taskId, ghSha, ghToken, hostingUrl, stoatConfigFileId);
+  const requestBody: UploadStaticHostingRequest = {
+    ghSha,
+    taskId,
+    stoatConfigFileId,
+    ghToken,
+    hostingUrl
+  };
+  await submitPartialConfig<UploadStaticHostingRequest>(taskId, requestBody);
 };
 
 export default runStaticHostingPlugin;
